Create Peak element
Given an array arr of n elements that is first strictly increasing and then maybe strictly decreasing, find the maximum element in the array.

If the first element is greater than the second or the last element is greater than the second last, print the respective element and terminate the program.
Else traverse the array from the second index to the second last index i.e. 1 to N – 1  
If for an element array[i] is greater than both its neighbors, i.e., array[i] > =array[i-1]  and array[i] > =array[i+1] , then print that element and terminate.